# openapi.yaml
openapi: 3.0.3
info:
  title: C2PA Soft Binding Resolution API
  description: >
    Implementation of the C2PA Soft Binding Resolution API
    (https://spec.c2pa.org/specifications/specifications/2.4/softbinding/Decoupled.html).
    Allows clients to store C2PA Manifest Stores, associate them with soft
    binding values (watermarks or fingerprints), and recover those manifests
    from a previously detected binding.
  version: '2.4'
servers:
  - url: /v1
security:
  - oauth2:
      - fetch:manifests
...
paths:
  /matches/byBinding:
    get:
      tags: [Query]
      summary: Find manifests by a soft binding value (query string)
      operationId: getMatchesByBinding
      parameters:
        - $ref: '#/components/parameters/value'
        - $ref: '#/components/parameters/alg'
        - $ref: '#/components/parameters/maxResults'
      responses:
        '200':
          $ref: '#/components/responses/Matches'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServiceFailure'
    post:
      tags: [Query]
      summary: Find manifests by a soft binding value (request body)
      description: Used when the binding value is too large to fit in a query string.
      operationId: postMatchesByBinding
      parameters:
        - $ref: '#/components/parameters/maxResults'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value, alg]
              properties:
                value:
                  type: string
                  description: Base64-encoded soft binding value
                alg:
                  type: string
                  description: Soft binding algorithm identifier
      responses:
        '200':
          $ref: '#/components/responses/Matches'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/ServiceFailure'
  /matches/byContent:
    post:
      tags: [Query]
      summary: Extract a soft binding from an uploaded asset and find matches
      operationId: getMatchesByContent
      parameters:
        - $ref: '#/components/parameters/algOptional'
        - $ref: '#/components/parameters/hintAlg'
        - $ref: '#/components/parameters/hintValue'
        - $ref: '#/components/parameters/maxResults'
      requestBody:
        required: true
        content:
          image/*:
            schema:
              type: string
              format: binary
          audio/*:
            schema:
              type: string
              format: binary
          video/*:
            schema:
              type: string
              format: binary
          application/*:
            schema:
              type: string
              format: binary
          model/*:
            schema:
              type: string
              format: binary
          text/*:
            schema:
              type: string
              format: binary
      responses:
        '200':
          $ref: '#/components/responses/Matches'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          description: Unsupported asset type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
  /matches/byReference:
    post:
      tags: [Query]
      summary: Download an asset by reference URL, extract a soft binding, and find matches
      operationId: getMatchesByReference
      parameters:
        - $ref: '#/components/parameters/algOptional'
        - $ref: '#/components/parameters/hintAlg'
        - $ref: '#/components/parameters/hintValue'
        - $ref: '#/components/parameters/maxResults'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [referenceUrl, assetLength]
              properties:
                referenceUrl:
                  type: string
                  format: uri
                  description: HTTPS URL the server will fetch the asset from
                assetLength:
                  type: integer
                  description: Expected size of the asset in bytes
                assetType:
                  type: string
                  description: Expected MIME type of the asset
                region:
                  description: Optional region of interest within the asset
      responses:
        '200':
          $ref: '#/components/responses/Matches'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '502':
          description: Failed to fetch the reference URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
...
  /manifests/{manifestId}:
    get:
      tags: [Fetch]
      summary: Retrieve a stored C2PA Manifest Store
      operationId: getManifest
      parameters:
        - $ref: '#/components/parameters/manifestId'
      responses:
        '200':
          description: The C2PA Manifest Store
          content:
            application/c2pa:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: C2PA Manifest not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
    delete:
      tags: [Store]
      summary: Remove a manifest and its associated bindings
      operationId: deleteManifest
      security:
        - oauth2:
            - store:manifests
      parameters:
        - $ref: '#/components/parameters/manifestId'
      responses:
        '204':
          description: Manifest deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: C2PA Manifest Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
  /manifests/{manifestId}/receipts:
    get:
      tags: [Fetch]
      summary: Retrieve the repository receipt for a manifest
      operationId: getManifestReceipt
      parameters:
        - $ref: '#/components/parameters/manifestId'
      responses:
        '200':
          description: The stored receipt, including verification status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: C2PA Manifest Store or receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
    post:
      tags: [Fetch]
      summary: Verify a caller-supplied receipt against a stored manifest
      operationId: verifyManifestReceipt
      parameters:
        - $ref: '#/components/parameters/manifestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Receipt'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '404':
          description: C2PA Manifest not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: OAuth2 client credentials. Configure the token URL for your identity provider.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/oauth/token
          scopes:
            fetch:manifests: Query and fetch manifests and receipts
            store:manifests: Store and delete manifests
            store:bindings: Create and update soft bindings
  parameters:
    value:
      name: value
      in: query
      required: true
      description: Base64-encoded soft binding value
      schema:
        type: string
    alg:
      name: alg
      in: query
      required: true
      description: Soft binding algorithm identifier
      schema:
        type: string
    algOptional:
      name: alg
      in: query
      required: false
      description: Soft binding algorithm to use for extraction
      schema:
        type: string
    hintAlg:
      name: hintAlg
      in: query
      required: false
      description: Algorithm identifier for a caller-supplied binding hint
      schema:
        type: string
    hintValue:
      name: hintValue
      in: query
      required: false
      description: Caller-supplied binding value used if extraction fails
      schema:
        type: string
    maxResults:
      name: maxResults
      in: query
      required: false
      description: Maximum number of matches to return
      schema:
        type: integer
        minimum: 1
        default: 10
    manifestId:
      name: manifestId
      in: path
      required: true
      description: Identifier of a stored C2PA Manifest Store
      schema:
        type: string
...
  responses:
    Matches:
      description: Matching manifests, ordered by relevance
      content:
        application/json:
          schema:
            type: object
            properties:
              matches:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The authenticated principal lacks the required authorization scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Request body exceeds the configured size limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceFailure:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Match:
      type: object
      properties:
        manifestId:
          type: string
        similarityScore:
          type: number
    Receipt:
      type: object
      properties:
        '@context':
          type: object
          properties:
            c2pa:
              type: string
            receipt:
              type: string
        '@type':
          type: string
          example: org.c2pa.manifest-receipt
        repository:
          type: object
          properties:
            uri:
              type: string
            manifestId:
              type: string
        anchor:
          type: object
          properties:
            uri:
              type: string
            proof:
              type: object
              properties:
                alg:
                  type: string
                value:
                  type: string
        verified:
          type: boolean
        error:
          type: string